+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [unreleased]
 
+- Remove func draw.Inline(), moved to design as private
 - Remove type Arrow, use Line
 
 ## [0.21.2] 2021-04-21